fwr: handle time zone definition and validation during the installation
git-svn-id: svn+ssh://kelev.kaltura.com/usr/local/kalsource/on-premise/trunk/installer@87150 6b8eccd3-e8c5-4e7d-8186-e12b5326b719

+++ installer/uninstall.php
@@ -96,13 +96,16 @@ if ((!$silentRun) && (!getTrueFalse(false))) {
 	die(0);
 }
 //We first want to clear the crontab symbolic link so we can remove all services without them returning
-foreach ($config['symlinks'] as $slink) {
-	echo 'Removing '.$slink.'... ';
-	if (execute('rm -rf ' . $slink)) {
-		echo 'OK'.PHP_EOL;
-	} else {
-		echo 'Failed'.PHP_EOL;
-		$success = false;
+if(is_array($config['symlinks']))
+{
+	foreach ($config['symlinks'] as $slink) {
+		echo 'Removing '.$slink.'... ';
+		if (execute('rm -rf ' . $slink)) {
+			echo 'OK'.PHP_EOL;
+		} else {
+			echo 'Failed'.PHP_EOL;
+			$success = false;
+		}
 	}
 }
